Open browser with xdg-open
On Linux, it's better to open web pages with `xdg-open` (opens with default programs based on MIME type, and installed by default on many distros) and let that figure out what the default browser is. In some distros, even with Chrome as default, it's `google-chrome-stable` rather than `google-chrome`, and many people probably don't use Chrome either.